requirements.txt
google-cloud-happybase==0.24.0
google-cloud-bigtable==0.24.0
google-cloud-core==0.24.1
